fix cio to same as cppio.

+++ iotest/cio.c
@@ -2,7 +2,7 @@
 
 int main()
 {
-        unsigned long i = 0;
+        unsigned int i = 0;
         for (; i < 0xFFFF; ++i)
                 printf("Hello\n");
         return 0;
